entity_tests/requirements.txt
treelib
pyserial
requests
ws4py
websocket-client
pytest
pytest-cov
virtualenv
